# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Pershing307 Member Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin) with optional pagination and search
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResult"
        "401":
          description: Not authenticated
        "403":
          description: Insufficient permissions
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user details (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResult"
        "404":
          description: User not found
  /users/{id}/deactivate:
    patch:
      operationId: deactivateUser
      tags: [users]
      summary: Deactivate a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/activate:
    patch:
      operationId: activateUser
      tags: [users]
      summary: Activate a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/roles:
    post:
      operationId: grantUserRole
      tags: [users]
      summary: Grant a role to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignment"
      responses:
        "200":
          description: Role granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/roles/{roleId}:
    delete:
      operationId: revokeUserRole
      tags: [users]
      summary: Revoke a role from a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Role revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/domains:
    get:
      operationId: getUserDomains
      tags: [users]
      summary: List domain access grants for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Domain access list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDomainAccessListResult"
    post:
      operationId: grantUserDomain
      tags: [users]
      summary: Grant a protected domain access to a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainGrantInput"
      responses:
        "200":
          description: Access granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/domains/{domainId}:
    delete:
      operationId: revokeUserDomain
      tags: [users]
      summary: Revoke protected domain access from a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: domainId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Access revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/degrees:
    get:
      operationId: getUserDegrees
      tags: [users]
      summary: List degree records for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Degree list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDegreeListResult"
    post:
      operationId: addUserDegree
      tags: [users]
      summary: Add a degree record for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DegreeInput"
      responses:
        "201":
          description: Degree added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
  /users/{id}/degrees/{degreeId}:
    delete:
      operationId: removeUserDegree
      tags: [users]
      summary: Remove a degree record from a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: degreeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Degree removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"
...
components:
  schemas:
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    UserSummary:
      type: object
      required: [id, email, firstName, lastName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        membershipStatus:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserListResult:
      type: object
      required: [users, total, limit, offset]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    RoleDetail:
      type: object
      required: [id, name, slug, permissionLevel]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        permissionLevel:
          type: integer
    UserDetail:
      type: object
      required: [id, email, firstName, lastName, isActive, emailVerified, mustChangePassword, createdAt, roles]
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: ["string", "null"]
        membershipStatus:
          type: ["string", "null"]
        isActive:
          type: boolean
        emailVerified:
          type: boolean
        mustChangePassword:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleDetail"
    UserDetailResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/UserDetail"
    RoleAssignment:
      type: object
      required: [roleId]
      properties:
        roleId:
          type: string
          minLength: 1
...
    UserDomainGrant:
      type: object
      required: [domainId, domainName, domainSlug, grantedAt]
      properties:
        domainId:
          type: string
        domainName:
          type: string
        domainSlug:
          type: string
        grantedAt:
          type: string
    UserDomainAccessListResult:
      type: object
      required: [domains]
      properties:
        domains:
          type: array
          items:
            $ref: "#/components/schemas/UserDomainGrant"
    DomainGrantInput:
      type: object
      required: [domainId]
      properties:
        domainId:
          type: string
          minLength: 1
    UserDegree:
      type: object
      required: [id, degree, degreeName, createdAt]
      properties:
        id:
          type: string
        degree:
          type: integer
        degreeName:
          type: string
        conferredOn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    UserDegreeListResult:
      type: object
      required: [degrees]
      properties:
        degrees:
          type: array
          items:
            $ref: "#/components/schemas/UserDegree"
    DegreeInput:
      type: object
      required: [degree]
      properties:
        degree:
          type: integer
          minimum: 1
        conferredOn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...