# lib/api-zod/src/generated/types/degreeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */

export interface DegreeInput {
  /** @minimum 1 */
  degree: number;
  /** @nullable */
  conferredOn?: string | null;
  /** @nullable */
  notes?: string | null;
}