# lib/api-zod/src/generated/types/userDegree.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pershing307 Member Portal API
 * OpenAPI spec version: 0.2.0
 */

export interface UserDegree {
  id: string;
  degree: number;
  degreeName: string;
  /** @nullable */
  conferredOn?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}